Ordner und Dateistruktur angepasst
